dom/html/test/test_htmlcollection.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=772869
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 772869</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=772869">Mozilla Bug 772869</a>
<p id="display"></p>
<div id="content" style="display: none">
  <a class="foo" name="x"></a>
  <span class="foo" id="y"></span>
  <span class="foo" name="x"></span>
  <form class="foo" name="z" id="w"></form>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 772869 **/
var x = document.getElementsByClassName("foo");
x.something = "another";
var names = [];
for (var name in x) {
  names.push(name);
}
is(names.length, 8, "Should have 8 enumerated names");
is(names[0], "0", "Enum entry 1")
is(names[1], "1", "Enum entry 2")
is(names[2], "2", "Enum entry 3")
is(names[3], "3", "Enum entry 4")
is(names[4], "something", "Enum entry 5")
is(names[5], "item", "Enum entry 6")
is(names[6], "namedItem", "Enum entry 7")
is(names[7], "length", "Enum entry 8");

names = Object.getOwnPropertyNames(x);
is(names.length, 9, "Should have 9 items");
is(names[0], "0", "Entry 1")
is(names[1], "1", "Entry 2")
is(names[2], "2", "Entry 3")
is(names[3], "3", "Entry 4")
is(names[4], "x", "Entry 5")
is(names[5], "y", "Entry 6")
is(names[6], "w", "Entry 7")
is(names[7], "z", "Entry 8")
is(names[8], "something", "Entry 9")
</script>
</pre>
</body>
</html>
